+++ saas-ui/charts/src/line-chart.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react'
 
 import {
-  chakra,
   Box,
   SystemProps,
   useColorModeValue,
@@ -17,10 +16,10 @@ import {
   CartesianGrid,
   Tooltip,
   ResponsiveContainer,
+  TooltipProps,
 } from 'recharts'
 
-/* @ts-ignore */
-import { DefaultTooltipContent } from 'recharts/lib/component/DefaultTooltipContent'
+import { ClassNames } from '@emotion/react'
 
 export interface ChartData {
   x: number
@@ -36,8 +35,10 @@ export interface LineChartProps {
   color?: string
   strokeWidth?: string
   name?: string
-  tickFormatter?: (value: number) => string
+  tickFormatter?(value: number): string
   variant?: 'line' | 'solid' | 'gradient'
+  tooltipContent?(props: TooltipProps<any, any>): React.ReactNode
+  tooltipFormatter?(value: string, name: string, props: any): string
 }
 
 export const LineChart = (props: LineChartProps) => {
@@ -50,6 +51,10 @@ export const LineChart = (props: LineChartProps) => {
     name,
     tickFormatter,
     variant,
+    tooltipContent,
+    tooltipFormatter = (value: string, name: string, props: any) => {
+      return props.payload.yv
+    },
   } = props
 
   const theme = useTheme()
@@ -69,64 +74,83 @@ export const LineChart = (props: LineChartProps) => {
     }
   })()
 
+  const tooltipStyles = css(styles)(theme)
+
   return (
-    <Box height={height}>
-      <ResponsiveContainer width="100%" height="100%">
-        <AreaChart
-          width={500}
-          height={300}
-          data={data}
-          margin={{
-            top: 0,
-            right: 0,
-            left: 0,
-            bottom: 0,
-          }}
-        >
-          <defs>
-            <linearGradient id="chart-gradient" x1="0" y1="0" x2="0" y2="1">
-              <stop offset="5%" stopColor={strokeColor} stopOpacity={0.8} />
-              <stop offset="95%" stopColor={strokeColor} stopOpacity={0} />
-            </linearGradient>
-          </defs>
-          {showGrid && (
-            <CartesianGrid
-              strokeDasharray=" 1 1 1"
-              vertical={false}
-              strokeOpacity={useColorModeValue(0.8, 0.3)}
-            />
-          )}
-          <XAxis dataKey="xv" />
-          <YAxis dataKey="y" tickFormatter={tickFormatter} />
-          <Tooltip
-            formatter={(value: string, name: string, props: any) => {
-              return props.payload.yv
-            }}
-            wrapperStyle={{ outline: 'none' }}
-            contentStyle={{
-              background: 'transparent',
-              border: '0',
-              padding: '0',
-              outline: 'none',
-            }}
-            content={(props) => {
-              return (
-                <chakra.div __css={styles}>
-                  <DefaultTooltipContent {...props} />
-                </chakra.div>
-              )
-            }}
-          />
-          <Area
-            type="monotone"
-            dataKey="y"
-            stroke={strokeColor}
-            strokeWidth={strokeWidth}
-            fill={fill}
-            name={name}
-          />
-        </AreaChart>
-      </ResponsiveContainer>
-    </Box>
+    <ClassNames>
+      {({ css }) => {
+        return (
+          <Box height={height}>
+            <ResponsiveContainer width="100%" height="100%">
+              <AreaChart
+                width={500}
+                height={300}
+                data={data}
+                margin={{
+                  top: 0,
+                  right: 0,
+                  left: 0,
+                  bottom: 0,
+                }}
+              >
+                <defs>
+                  <linearGradient
+                    id="chart-gradient"
+                    x1="0"
+                    y1="0"
+                    x2="0"
+                    y2="1"
+                  >
+                    <stop
+                      offset="5%"
+                      stopColor={strokeColor}
+                      stopOpacity={0.8}
+                    />
+                    <stop
+                      offset="95%"
+                      stopColor={strokeColor}
+                      stopOpacity={0}
+                    />
+                  </linearGradient>
+                </defs>
+                {showGrid && (
+                  <CartesianGrid
+                    strokeDasharray=" 1 1 1"
+                    vertical={false}
+                    strokeOpacity={useColorModeValue(0.8, 0.3)}
+                  />
+                )}
+                <XAxis dataKey="xv" />
+                <YAxis dataKey="y" tickFormatter={tickFormatter} />
+
+                <Tooltip
+                  formatter={tooltipFormatter}
+                  wrapperStyle={{ outline: 'none' }}
+                  contentStyle={{
+                    background: 'var(--tooltip-bg)',
+                    border:
+                      '1px solid var(--chakra-colors-default-border-color)',
+                    outline: 'none',
+                    display: 'block',
+                    padding: '4px 8px',
+                  }}
+                  wrapperClassName={css(tooltipStyles)}
+                  content={tooltipContent}
+                />
+
+                <Area
+                  type="monotone"
+                  dataKey="y"
+                  stroke={strokeColor}
+                  strokeWidth={strokeWidth}
+                  fill={fill}
+                  name={name}
+                />
+              </AreaChart>
+            </ResponsiveContainer>
+          </Box>
+        )
+      }}
+    </ClassNames>
   )
 }